pyproject.toml
[project]
name = "gitlab-docs"
[project.urls]
"Home Page" = "https://pypi.org/manage/project/gitlab-docs"
DOCUMENTATION = "https://pypi.org/manage/project/gitlab-docs"
Repository = "https://gitlab.com/charlieasmith/gitlab-docs"
Gitlab     = "https://gitlab.com/charlieasmith/gitlab-docs"

[tool.poetry_bumpversion.file."gitlab_docs/__init__.py"]
[tool.poetry]
name = "gitlab-docs"
version = "0.1.1"
description = "A tool that automatically generates gitlab documentation from yaml files"
authors = ["Charlie Smith <me@charlieasmith.co.uk>"]
readme = "README.md"
packages = [{include = "gitlab_docs", format = "sdist"}]

[tool.poetry.dependencies]
python = ">=3.12"
prettytable = "*"
pytablewriter = "~1.2.0"
oyaml = "*"
typer = "^0.10.0"
markdown-analysis = "*"
setuptools = "*"
jq = "^1.8.0"
semver = "^3.0.2"
click = "<=7.1.2"



[tool.poetry.scripts]
gitlab-docs = 'gitlab_docs.app:gitlab_docs'

[tool.poetry.group.dev.dependencies]
md-click = "^1.0.1"

[build-system]
requires = ["poetry-core","setuptools","poetry-bumpversion","id"]
build-backend = "poetry.core.masonry.api"
